move graphs into separate file
